requirements.txt
rPI.GPIO
flask
flask-socketio
